moving files into correct directory
