+++ profiles/forms.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 from django.forms import ModelForm
 from django.utils.translation import ugettext_lazy as _
 
-from profiles.models import Address, Membership, Profile
+from profiles.profiles.models import Address, Membership, Profile
 
 
 class PasswordChangingForm(PasswordChangeForm):
