+++ profiles/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import views
+from profiles.profiles import views
 
 urlpatterns = [
     path('my_profile', views.my_profile, name='my_profile'),
